ex02 done

+++ day07/ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: ebelkhei <ebelkhei@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/18 11:22:08 by ebelkhei          #+#    #+#             */
-/*   Updated: 2023/05/18 11:43:26 by ebelkhei         ###   ########.fr       */
+/*   Updated: 2023/05/18 12:11:53 by ebelkhei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@
 #include <string>
 
 template <typename T>
-void iter(T *arr, unsigned int len, void (*f)(T))
+void iter(T *arr, size_t len, void (*f)(T))
 {
     if (!arr || !f)
         return ;
 
-    for (unsigned int i = 0; i < len; i++)
+    for (size_t i = 0; i < len; i++)
         f(arr[i]);
 }
